// src/app/types/types.ts
const phoneNumberRegex = /^[+]?[1-9]\d{1,14}$/;
export interface User {
  id: string; 
  userName: string; 
  email: string;
  password: string;
  createdAt?: Date|null
  updatedAt?: Date |null
  news: News[]
  role: string
}
export interface UpdateUser {
  id: string; 
  userName: string; 
  email: string;
  password: string;
  role: string

}


export const doctorZodSchema = z.object({
  img:z.instanceof(File).optional().nullable(),
  doctorName: z.string().min(10, "اسم الدكتور يجب أن يكون على الأقل 10 أحرف"),
  phone: z.string().regex(phoneNumberRegex, "رقم الهاتف يجب أن يتكون من 9 أرقام"),
  specialist: z.string().min(3, "التخصص يجب أن يكون على الأقل 3 أحرف"),
  department: z.object({
    depName: z.string().min(2,"اسم القسم مطلوب"),
  }),// يمكنك تحديد ما إذا كان هذا الحقل مطلوبًا أم لا
  weekwork: z.array(z.object({
    day: z.string().nonempty("اليوم مطلوب"),
    startTime: z.string().min(1,"وقت بدءالدوام ونهابة الدوام مطلوب"),
    endTime: z.string().min(1,"وقت بدءالدوام ونهابة الدوام مطلوب"),
  })).min(1," يجب إضافة أيام الدوام "), // يمكنك تحديد ما إذا كان هذا الحقل مطلوبًا أم لا
  information:z.string().optional()
});
export type doctroType = z.infer<typeof doctorZodSchema>;